+++ app/MainAppComponent.js
@@ -10,13 +10,13 @@ import {
   } from "react-router-dom";
 
   // navit
-import {default as DisplayCourseComponent} from "./DisplayCourseComponent";
-import {default as RosterDetails} from "./DisplaySavedRosterComponent";
-import {default as DisplayPickedFile} from "./DisplayPickedFile";
+import {default as DisplayCourseComponent} from "./Course/DisplayCourseComponent";
+import {default as RosterDetails} from "./Roster/DisplaySavedRosterComponent";
+import {default as DisplayPickedFile} from "./Roster/DisplayPickedFile";
 import { Table, Header, Cell, Icon, Button, Grid, GridColumn, Container } from 'semantic-ui-react';
 import {default as SidebarComponent } from "./SidebarComponent";
 import { default as HeaderComponent } from "./AppHeaderComponent";
-import { default as SurveyFormQuestionComponent } from "./SurveyFormQuestionComponent";
+import { default as SurveyFormQuestionComponent } from "./Survey/SurveyFormQuestionComponent";
 
 import {default as AnswerSurveyComponent } from "./StudentAnswerSurvey/AnswerSurveyComponent.jsx"; // rahul
